Move datastructure files into separate directory
