+++ src/interpreter/InterpreterBTreeIndex.cpp
@@ -15,9 +15,9 @@
  ***********************************************************************/
 
 #include "interpreter/InterpreterIndex.h"
-#include "souffle/BTree.h"
 #include "souffle/CompiledTuple.h"
 #include "souffle/RamTypes.h"
+#include "souffle/datastructure/BTree.h"
 #include "souffle/utility/MiscUtil.h"
 #include <algorithm>
 #include <cstddef>
